Arbitrary bit sizes for slots in the counting bloom filter
Relax the divisible-by restriction on the number of bits for each slot.
I suspect the sweet spot for these things will be 3 bits, so this change
seemed worth doing. It was trickier than expected. I created a
"visitSlotBytes" method to refactor incrementSlot and getSlot to avoid
duplicated code as it had gotten complicated.

Fixed visibility issues with filter class.