Removed helper methods on gossip for adding and getting info objects of the basic types. This was just a lot of extra lines of code without much of a purpose.
Summary:
Added support for arbitrary types of info objects, so long as they
implement the gossip.Ordered interface. This will allow us to add
more complicated struct types to groups, something we'll need to do
virtually everywhere. For example, when keeping track of disks with
available capacity per datacenter, we're going to want to store a
struct which describes the location of the disk host, port, device
and also contains more detailed information about free space. We
could previously have stored this all in the key, but not worth
having to parse it out!

Test Plan: unittests included

Reviewers: andybons, levonlloyd

Reviewed By: andybons, levonlloyd

Differential Revision: http://phabricator.andybons.com/D16