Integrated a CLI interface to the main cockroach binary which currently has two commands: init & start.
Summary:
A new cluster is init'd by specifying a single data directory as first
argument and a default zone configuration file as second arg. A new
RocksDB instance is created in the data directory (an error results if
one is already there), and a new cluster UUID is generated to uniquely
identify the cockroach cluster. The first range is created as a single
replica with ident: NodeID: 1, StoreID: 1, RangeID: 1. meta1/meta2
records are created for the first range, and node id sequence generator,
store id sequence generator, and store-local range id sequence generator
are all created.

The roach node is run using "cockroach start". Typically this will be
accompanied by --bootstrap_hosts in order to connect to gossip network,
and --data_dirs=dir1,dir2,...  The --data_dirs can either have a list
of paths or a list of "mem=100000,mem=200000,...", which creates in-mem
stores. Any stores which are as yet uninitialized are bootstrapped
using the cluster ID and new store IDs from the store id sequence
generator. Stores other than the one used to init a new cluster,
or all stores on a new node being added to an existing cluster will need
initialization.

Changed Replica specification to be a series of three integers: NodeID,
StoreID & RangeID. Node & store IDs are 32-bit; range ID is 64-bit.

Added convenience methods to kv.DB as well as storage.Engine for reading
and writing arbitrary go values via interface{}.

Test Plan: go test

Reviewers: petermattis, andybons, bdarnell

Reviewed By: bdarnell

Subscribers: hrsht, levonlloyd

Differential Revision: http://phabricator.andybons.com/D29