RocksDB get/put/del and appropriate vendoring.
Summary:
This change adds a Makefile and _vendor directory for dealing with cgo dependencies.
The details of how vendoring RocksDB will work is outlined within _vendor/README.md.

The RocksDB subtree is based on our own fork, so our changes will be brought in
appropriately. It seemed like overkill to create our own gorocks library and was
becoming a major hassle to deal with the build flow, which is why I’m eschewing
that repo for now. If this fits our needs, then I don’t think we need to have
another repo for it lying around.

Test Plan: make test

Reviewers: spencerkimball, bdarnell

Reviewed By: bdarnell

Subscribers: bdarnell, strangemonad, team

Differential Revision: http://phabricator.andybons.com/D47