Removed dial timeout from gossip.Client. Instead, let the rpc.Client retry indefinitely. Meanwhile, the gossip.Gossip instance will wait for its checkTimeout to fire and notice the gossip network hasn't been connected and so fire off its next...
Summary:
...boostrap attempt.

rpc.Client now insists on at least one successful heartbeat before
signalling that the connection is ready.

Souped up the retrywithbackoff method to take a "tag" in the options
which describes the point of the retry loop and is included in log
statements for clarity.

Test Plan: go test

Reviewers: Tobias, andybons

Reviewed By: andybons

Differential Revision: http://phabricator.cockroachdb.org/D59