add test to verify that redundant gossip clients are dropped
Summary: looked at this because the simulation sometimes shows a persistent connection loop between two gossip nodes (i.e. a connects to b and b connects to a). It seems that the detection for this is working alright though. Also the breakup is definitely triggered at times during the simulation. So I'm not wiser than before but we have one more test for this sort of thing.

Test Plan: make test

Reviewers: andybons, spencerkimball

Reviewed By: spencerkimball

Differential Revision: http://phabricator.cockroachdb.org/D81