Modification to RangeDescriptor
Summary:
This is a modification to `RangeDescriptor`, as part of ongoing work to improve
range metadata lookup.

Previously, `RangeDescriptor` contained a list of replicas, and the Metadata
StartKey for the range (corresponding to the Metadata EndKey, the key at which
the RangeDescriptor is stored). The `RangeDescriptor` now contains the replica,
along with the Start and End keys describing the range of keys actually
contained by the range.

+ Added method `MetadataKey()`, which returns an appopriate metadata key
corresponding to a given key.
+ `RangeDescriptor` is now an anonymous field of `RangeMetadata`.  RangeMetadata
no longer contains its own `StartKey` and `EndKey` fields.

Test Plan: go test

Reviewers: Tobias, andybons, spencerkimball

Reviewed By: spencerkimball

Subscribers: team

Differential Revision: http://phabricator.cockroachdb.org/D120