Add the definition for Transaction. Added a transaction coordinator in DistDB to handle the heartbeat message for clients.
Summary:
This diff added the definition for Transaction. This is the value type to be stored in the transaction table.
Also the stub for BeginTransaction is added.
Another part is the transaction coordinator in DistDB. Each DistDB has a transaction coordinator instance. The transaction
coordinator will send heartbeat messages to the transaction table for each transaction, so that other transactions know
whether this transaction is alive. This diff implemented the heartbeat message part.

Test Plan: go test

Reviewers: jiangmingyang, spencerkimball

Reviewed By: spencerkimball

Subscribers: mrtracy, team

Differential Revision: http://phabricator.cockroachdb.org/D118