Optionally use TLS for RPC
Summary:
TLS can be enabled by specifying the "-certs" flag to "cockroach start",
along with a directory containing a CA cert, node cert, and node key.
Samples of these files are in "resources/test_certs".

Test Plan:
The following seem to work:

    make test

    ./cockroach init hdd=test_space/node1
    ./cockroach start --gossip=localhost:8081 -rpc=localhost:8081 \
                -http=localhost:8080 -certs=resources/test_certs \
                -stores=hdd=test_space/node1

    # In a separate terminal:
    ./cockroach start --gossip=localhost:8081 -rpc=localhost:8281 \
                -http=localhost:8280 -certs=resources/test_certs \
                -stores=hdd=test_space/node2

    # Also the above, without the "-cert"s.

I haven't tested on separate machines yet -- note that that would require
different certs, since the ones in test_certs are for localhost.

I also haven't tried adding any data to the DB; is there a way to do that
on the command line?

Reviewers: spencerkimball

Reviewed By: spencerkimball

Subscribers: andybons, team, maniksurtani, bdarnell

Differential Revision: http://phabricator.cockroachdb.org/D125