Fix chicken/egg bug when bootstrapping.
Summary:
cockroach init was panicking due to a null rdb instance during bootstrap.
This was due to the fact that creation of the underlying RocksDB DB occurs in Init,
which will return an error if the store is not bootstrapped.

To allow for bootstrapping, Start the engine at both Init and Bootstrap but make
RocksDB.Start() idempotent after the first call.

question: Should this be made more explicit in the Engine interface definition?

Test Plan: unit tests pass

Reviewers: petermattis, spencerkimball

Reviewed By: spencerkimball

Subscribers: team

Differential Revision: http://phabricator.cockroachdb.org/D159