Fix filenames in log files.
The indirection here was causing all log lines to be reported as "log.go".
Using the glog functions more directly lets the original filename through,
at the expense of customizability.  When we make the changes that motivated
the introduction of the log package, we'll need to either revert this or
come up with another solution to preserve the caller's file/line info.