Addressed all PR comments.
Added proto.Response.Verify method to verify checksums and now do verification
automatically on the responses of Get(), ConditionalPut() and Scan(). Also
now always initialize checksums on Put() and ConditionalPut().

Eventually we're going to want to fully expose the checksums to non-go
clients, but will have to evolve the KV API to include that capability.