Implement and unittest EndTransaction.
Added an implementation for EndTransaction which attempts to move a
transaction from PENDING to either COMMITTED or ABORTED depending on
the EndTransactionRequest.Commit argument. Various error conditions
may result, including the txn already being committed or aborted,
Epoch or timestamp regressions, or transaction retry on SSI txns
in the event that the commit timestamp is different from the transaction
timestamp.