Added SQL parser based on vitess's SQL parser.
Local changes include support for a bit more of the SQL language and
some naming changes to make this package be almost lint free. Note that
there is a remaining bit of lint in the generated sql.go file that I
cannot get rid of because it is generated by yacc (yyEofCode should be
named yyEOFCode).