Last clock offset cleanup.
Limit of 1.5 seconds for measuring clock offset,
to limit the maximum error on a read.

The remote clock reading with smallest error since the last cluster
offset measurement is now used, rather than the latest reading.
This makes it more likely to get a good measurement before we
determine the offset interval.