add testability through docker
added a ./deploy/wrapper.sh which routes everything to the cockroach binary
except when called when with a single parameter "test", in which case we run
make test and make testrace.

Somewhat optimized the size of the resulting Docker image and set the entry
point to the new wrapper script, allowing
  `docker run "cockroachdb/cockroach" test
to run the tests inside of the container.