setup CircleCI to test through Docker with caching
The script will look for previously generated Docker images and use those in
the build of our Cockroach image. On the first run (or after the cache has been
cleared) we do a full build and store the layers.
In effect, this gets the total build time down to about two minutes, where half
of that is due to the need to run `docker load` on our currently 1.2GB sized
container.