Added a grace interval to split_test when verifying there are no dangling intents. This allows the test writer to finish up and also any ongoing split to complete.
Removed useless test from gossip/addr_set_test which could occasionally
fail after selecting the 10 addresses with no overlaps in 10 tries. The
test originally was predicated on a consistent seed in random number
generator, but that's apparently no longer the case with the unittests.
It wasn't a test worth keeping, though we may want to look into our test
random number generator so that at least we are logging the random seed.