Move heartbeat and gossip request/responses to protobuf.
This is anticipation of using other RPC codecs or an entirely different
transport (HTTP/2).

The remaining "icky" bit of this is that infostore deltas are still
encoded using gob. We're going to want to create a protobuf description
of infostore deltas.

I had originally included a protobuf codec but it felt very iffy since
the tests were weak, etc. I feel safer with gob still.